Add ESLint rules for linting
