+++ backend/dashboardd/.eslintrc.yml
@@ -1,0 +1,11 @@
+extends: airbnb
+plugins:
+  - react
+  - jsx-a11y
+  - import
+env:
+  browser: true
+settings:
+  import/resolver:
+    webpack:
+      config: 'config/webpack.config.dev.js'
