+++ backend/dashboardd/config/paths.js
@@ -68,7 +68,7 @@ module.exports = {
   appBuild: resolveApp('../../bin/dashboard'),
   appPublic: resolveApp('public'),
   appHtml: resolveApp('public/index.html'),
-  appIndexJs: resolveApp('src/index.js'),
+  appIndexJs: resolveApp('src/index.jsx'),
   appPackageJson: resolveApp('package.json'),
   appSrc: resolveApp('src'),
   yarnLockFile: resolveApp('yarn.lock'),
