+++ backend/dashboardd/src/containers/app/index.js
@@ -1,3 +1,3 @@
-import App from './app'
+import App from './app';
 
-export default App
+export default App;
