+++ backend/dashboardd/src/containers/events/index.js
@@ -1,3 +1,3 @@
-import Events from './events'
+import Events from './events';
 
-export default Events
+export default Events;
