+++ backend/dashboardd/src/index.jsx
@@ -4,18 +4,19 @@ import injectTapEventPlugin from 'react-tap-event-plugin';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import App from 'containers/app';
-import Events from 'containers/events'
+import Events from 'containers/events';
 
-import 'typeface-roboto'
-import './index.css'
+import './index.css';
+
+require('typeface-roboto');
 
 injectTapEventPlugin();
 
 ReactDOM.render(
   <Router>
     <App>
-      <Route exact path="/" component={Events}/>
+      <Route exact path="/" component={Events} />
     </App>
   </Router>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
